+++ packages/graphql/src/types.ts
@@ -21,6 +21,7 @@ import type { EventEmitter } from "events";
 import type { InputValueDefinitionNode, DirectiveNode, TypeNode, GraphQLSchema } from "graphql";
 import type { ResolveTree } from "graphql-parse-resolve-info";
 import type { Driver, Integer, Session, Transaction } from "neo4j-driver";
+import type Cypher from "@neo4j/cypher-builder";
 import type { Node, Relationship } from "./classes";
 import type { Neo4jDatabaseInfo } from "./classes/Neo4jDatabaseInfo";
 import type { RelationshipQueryDirectionOption } from "./constants";
@@ -368,7 +369,7 @@ export interface Neo4jGraphQLAuthPlugin {
     /**
      * This function tries to resolve public or secret keys.
      * The implementation on how to resolve the keys by the `JWKSEndpoint` or by the `Secret` is set on when the plugin is being initiated.
-     * @param req 
+     * @param req
      */
     tryToResolveKeys(req: unknown): void;
 }
@@ -524,3 +525,10 @@ export interface Neo4jFiltersSettings {
 export interface Neo4jFeaturesSettings {
     filters?: Neo4jFiltersSettings;
 }
+
+export type PredicateReturn = {
+    predicate: Cypher.Predicate | undefined;
+    preComputedSubqueries?: Cypher.CompositeClause | undefined;
+    requiredVariables: Cypher.Variable[];
+    aggregatingVariables: Cypher.Variable[];
+};
